+++ jac/jaclang/compiler/unitree.py
@@ -1388,13 +1388,9 @@ class ModuleItem(AstSymbolNode):
     @property
     def from_parent(self) -> Import:
         """Get import parent."""
-        if (
-            not self.parent
-            or not self.parent.parent
-            or not isinstance(self.parent.parent, Import)
-        ):
+        if not self.parent or not isinstance(self.parent, Import):
             raise ValueError("Import parent not found. Not Possible.")
-        return self.parent.parent
+        return self.parent
 
     @property
     def from_mod_path(self) -> ModulePath:
@@ -1533,7 +1529,8 @@ class Archetype(
                     new_kid.append(stmt)
                 new_kid.append(self.gen_token(Tok.RBRACE))
         else:
-            new_kid.append(self.gen_token(Tok.SEMI))
+            new_kid.append(self.gen_token(Tok.LBRACE))
+            new_kid.append(self.gen_token(Tok.RBRACE))
         self.set_kids(nodes=new_kid)
         return res
 
@@ -4664,7 +4661,7 @@ class String(Literal):
 
     def unparse(self) -> str:
         super().unparse()
-        return repr(self.value)
+        return self.value
 
 
 class Bool(Literal):