+++ scripts/tests.sh
@@ -1,5 +1,5 @@
 pytest -n 30 jac
 pytest jac-splice-orc
 pytest jac-cloud
-pytest jac-mtllm
+pytest jac-byllm
 jac test jac/examples/littleX/littleX.test.jac