+++ docs/docs/communityhub/release_notes.md
@@ -2,9 +2,13 @@
 
 This document provides a summary of new features, improvements, and bug fixes in each version of Jac and Jaseci. For details on changes that might require updates to your existing code, please refer to the [Breaking Changes](./breaking_changes.md) page.
 
-## jaclang 0.8.6 / jac-cloud 0.2.6 / mtllm 0.3.11 (Unreleased)
+## jaclang 0.8.7 / jac-cloud 0.2.7 / byllm 0.4.2 (Unreleased)
 
-## jaclang 0.8.5 / jac-cloud 0.2.5 / mtllm 0.3.10
+## jaclang 0.8.6 / jac-cloud 0.2.6 / byllm 0.4.1
+
+- **byLLM transition**: MTLLM has been transitioned to byLLM and PyPi package is renamed to `byllm`. Github actions are changed to push byllm PyPi. Alongside an mtllm PyPi will be pushed which installs latest `byllm` and produces a deprecation warning when imported as `mtllm`.
+
+## jaclang 0.8.5 / jac-cloud 0.2.5 / mtllm 0.4.0
 
 - **Jac Cloud Hot Reload**: Introduced the ability to enable development mode like uvicorn by adding `--reload` in `jac serve`. This supports targetting specific directories by using `--watch path/to/dir1,path/to/dir2` (comma separated).
 - **Dynamic Runtime Walker Endpoint**: Fixes auto-generated endpoints for walkers created at runtime.